test refactor

+++ back-end/tons-of-tacos/src/main/java/com/adamstraub/tonsoftacos/dto/businessDto/OrderItemReturnedToOwner.java
@@ -2,11 +2,14 @@ package com.adamstraub.tonsoftacos.dto.businessDto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class OrderItemReturnedToOwner {
     private Integer orderItemId;
     String itemName;
     private Integer quantity;
-    private Double total;
+//    private Double total;
+    private BigDecimal total;
 
 }
